# review-service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
    /v1/review:
        post:
            tags:
                - Review
            description: 创建评价
            operationId: Review_CreateReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.CreateReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.CreateReviewReply'
    /v1/review/appeal:
        post:
            tags:
                - Review
            description: 商家申诉评价
            operationId: Review_AppealReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AppealReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AppealReviewReply'
    /v1/review/appeal/audit:
        post:
            tags:
                - Review
            description: 运营审核申诉
            operationId: Review_AuditAppeal
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditAppealRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditAppealReply'
    /v1/review/audit:
        post:
            tags:
                - Review
            description: 运营审核评价
            operationId: Review_AuditReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.AuditReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.AuditReviewReply'
    /v1/review/detail/{reviewID}:
        get:
            tags:
                - Review
            description: 获取单条评价（路径避免与 /v1/review/order、/v1/review/user 冲突）
            operationId: Review_GetReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.GetReviewReply'
    /v1/review/order/{orderID}:
        get:
            tags:
                - Review
            description: 按订单分页列评价
            operationId: Review_ListReview
            parameters:
                - name: orderID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ListReviewReply'
    /v1/review/reply:
        post:
            tags:
                - Review
            description: 商家回复评价
            operationId: Review_ReplyReview
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.ReplyReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ReplyReviewReply'
    /v1/review/user/{userID}:
        get:
            tags:
                - Review
            description: 按用户分页列评价
            operationId: Review_ListReviewByUseId
            parameters:
                - name: userID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.ListReviewReply'
    /v1/review/{reviewID}:
        put:
            tags:
                - Review
            description: 修改评价
            operationId: Review_UpdateReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.review.v1.UpdateReviewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.UpdateReviewReply'
        delete:
            tags:
                - Review
            description: 删除评价
            operationId: Review_DeleteReview
            parameters:
                - name: reviewID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: userID
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.review.v1.DeleteReviewReply'
components:
    schemas:
        api.review.v1.AppealReviewReply:
            type: object
            properties:
                appealID:
                    type: string
        api.review.v1.AppealReviewRequest:
            type: object
            properties:
                userID:
                    type: string
                reviewID:
                    type: string
                reason:
                    type: string
                picInfo:
                    type: string
        api.review.v1.AuditAppealReply:
            type: object
            properties: {}
        api.review.v1.AuditAppealRequest:
            type: object
            properties:
                appealID:
                    type: string
                result:
                    type: integer
                    description: 1=通过 2=驳回
                    format: int32
                remark:
                    type: string
                operator:
                    type: string
        api.review.v1.AuditReviewReply:
            type: object
            properties:
                reviewID:
                    type: string
                status:
                    type: integer
                    format: int32
        api.review.v1.AuditReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                result:
                    type: integer
                    description: 1=通过(置为审核通过) 2=驳回(置为审核不通过)
                    format: int32
                remark:
                    type: string
                operator:
                    type: string
        api.review.v1.CreateReviewReply:
            type: object
            properties:
                reviewID:
                    type: string
        api.review.v1.CreateReviewRequest:
            type: object
            properties:
                userID:
                    type: string
                orderID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                anonymous:
                    type: boolean
        api.review.v1.DeleteReviewReply:
            type: object
            properties: {}
        api.review.v1.GetReviewReply:
            type: object
            properties:
                item:
                    $ref: '#/components/schemas/api.review.v1.ReviewListItem'
        api.review.v1.ListReviewReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.review.v1.ReviewListItem'
                total:
                    type: string
        api.review.v1.ReplyReviewReply:
            type: object
            properties:
                replyID:
                    type: string
        api.review.v1.ReplyReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                storeID:
                    type: string
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                extJSON:
                    type: string
                ctrlJSON:
                    type: string
        api.review.v1.ReviewListItem:
            type: object
            properties:
                reviewID:
                    type: string
                userID:
                    type: string
                orderID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
                status:
                    type: integer
                    format: int32
                hasReply:
                    type: integer
                    format: int32
                createAt:
                    type: string
        api.review.v1.UpdateReviewReply:
            type: object
            properties: {}
        api.review.v1.UpdateReviewRequest:
            type: object
            properties:
                reviewID:
                    type: string
                userID:
                    type: string
                score:
                    type: integer
                    format: int32
                serviceScore:
                    type: integer
                    format: int32
                expressScore:
                    type: integer
                    format: int32
                content:
                    type: string
                picInfo:
                    type: string
                videoInfo:
                    type: string
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Greeter
      description: The greeting service definition.
    - name: Review
      description: 评价服务